refactor is functional, still requires testing and more features

+++ recipes/server.rb
@@ -17,8 +17,107 @@
 # limitations under the License.
 #
 
-UNMANAGED_HOSTS_DATABAG = 'rdiff-backup_unmanagedhosts'
+HOSTS_DATABAG = 'rdiff-backup_hosts'
 CRON_FILE = '/etc/cron.d/rdiff-backup'
+
+# Recursive copy for objects like nested Hashes.
+def deep_copy(object)
+  Marshal.load(Marshal.dump(object))
+end
+
+# Recursive copy for Chef node hashes; only copies stuff relevant to rdiff-backup.
+def deep_copy_node(oldhash)
+  newhash = Hash.new
+  newhash['rdiff-backup'] = JSON.parse(oldhash['rdiff-backup'].to_json) # This is passed through JSON to convert it from an ImmutableMash to a plain old Hash.
+  newhash['fqdn'] = oldhash['fqdn']
+  newhash['chef_environment'] = oldhash['chef_environment']
+  return newhash
+end
+
+# Recursive bot.merge!(top) for hashes.
+def deep_merge!(bot, top)
+  top.keys.each do |key|
+    if top[key].is_a?(Hash)
+      unless bot[key].is_a?(Hash) # Make a new subhash in bot if top has one there.
+        bot[key] = Hash.new
+      end
+      deep_merge!(bot[key], top[key])
+    else
+      bot[key] = top[key]
+    end
+  end
+end
+
+servernode = deep_copy_node(node.to_hash); # This servernode hash that emulates a node definition allows rdiff-backup server attributes to be stored and modified without affecting the actual node definition.
+
+nodes = Set.new
+
+# Find nodes to back up by searching.
+Chef::Log.info("Beginning search for nodes. This may take some time depending on your node count.")
+searchhosts = search(:node, 'recipes:rdiff-backup\:\:client')
+searchhosts.each do |host|
+  hosthash = deep_copy_node(host.to_hash) # Convert the node to a deep hash for easy management.
+
+  # Create a new "node" hash for each host.
+  newnode = deep_copy_node(servernode) # Use the rdiff-backup server's attributes as the defaults.
+  deep_merge!(newnode, hosthash) # Override the defaults with the client node's attributes.
+
+  nodes << newnode # Add the new node to the set of nodes.
+end
+
+# Find nodes to back up from the hosts databag too.
+baghosts = data_bag(HOSTS_DATABAG).to_set
+
+# First pass to get the rdiff-backup server's databag first so we can use it for defaults if it exists.
+baghosts.each do |bagitem|
+  databagnode = data_bag_item(HOSTS_DATABAG, bagitem) # Read a "node" from the databag.
+  databagnode['fqdn'] = databagnode['id'].gsub('_', '.') # Fix the fqdn, since periods couldn't be used in the databag ID.
+  databagnode.delete('id')
+  if servernode['fqdn'] == databagnode['fqdn']
+    deep_merge!(servernode, databagnode) # Override the server's current attributes with attributes from its databag.
+    break
+  end
+end
+
+# Second pass for all the other hosts.
+baghosts.each do |bagitem|
+  databagnode = data_bag_item(HOSTS_DATABAG, bagitem) # Read a "node" from the databag.
+  databagnode['fqdn'] = databagnode['id'].gsub('_', '.') # Fix the fqdn, since periods couldn't be used in the databag ID.
+  databagnode.delete('id')
+  
+  # If the host is already being managed in Chef, use its node attributes as the defaults.
+  nodes.each do |n|
+    if databagnode['fqdn'] == n['fqdn']
+      clientnode = n
+      break
+    end
+  end
+
+  # Otherwise, use the rdiff-backup server's attributes as the defaults.
+  clientnode = deep_copy_node(servernode) unless clientnode
+
+  deep_merge!(clientnode, databagnode) # Override the defaults with the databag attributes.
+
+  nodes << clientnode # Add the new node to the set of nodes.
+end
+
+# Filter out clients not in our environment, if applicable.
+if servernode['rdiff-backup']['server']['restrict-to-own-environment']
+  nodes.each do |n|
+    if n['chef_environment'] != servernode['chef_environment']
+      nodes.delete(n)
+    end
+  end
+end
+
+# Filter out ourself, since the server should not also be a client.
+nodes.delete_if do |n|
+  n['fqdn'] == servernode['fqdn']
+end
+
+if nodes.empty?
+  Chef::Log.info("WARNING: No nodes returned from search or found in rdiff-backup_hosts databag.")
+end
 
 # Install rdiff-backup.
 package "rdiff-backup" do
@@ -26,74 +125,67 @@ package "rdiff-backup" do
 end
 
 # Create the server backup group.
-group node['rdiff-backup']['server']['user'] do
+group servernode['rdiff-backup']['server']['user'] do
   system true
 end
 
 # Create the server backup user.
-user node['rdiff-backup']['server']['user'] do
+user servernode['rdiff-backup']['server']['user'] do
   comment 'User for rdiff-backup server backups'
-  gid node['rdiff-backup']['server']['user']
+  gid servernode['rdiff-backup']['server']['user']
   system true
   shell '/bin/bash'
-  home '/home/' + node['rdiff-backup']['server']['user']
+  home '/home/' + servernode['rdiff-backup']['server']['user']
   supports :manage_home => true
 end
 
-# Note: The server backup user's private key must be copied over manually.
+# Copy over and set up the Nagios nrpe plugin, if applicable.
+if servernode['rdiff-backup']['server']['nagios']['alerts']
 
-nodes = Set.new
-
-# Find nodes to back up by searching.
-Chef::Log.info("Beginning search for nodes. This may take some time depending on your node count.")
-hosts = search(:node, 'recipes:rdiff-backup\:\:client')
-hosts.each do |host|
-  host = host.to_hash # Convert the node to a deep hash for easy management.
-
-  # Create a new "node" hash for each unmanaged host.
-  newnode = deep_copy(node) # Use the rdiff-backup server's attributes as the defaults.
-  deep_merge!(newnode, host) # Override the defaults with the client node's attributes.
-
-  nodes << newnode # Add the new node to the set of nodes.
-end
-
-# Find nodes to back up from the unmanagedhosts databag too.
-unmanagedhosts = data_bag(UNMANAGED_HOSTS_DATABAG).to_set
-unmanagedhosts.each do |bagitem|
-  host = data_bag_item(UNMANAGED_HOSTS_DATABAG, bagitem) # Read a "node" from the databag.
-  
-  # Create a new "node" hash for each unmanaged host.
-  newnode = deep_copy(node) # Use the rdiff-backup server's attributes as the defaults.
-  deep_merge!(newnode, host) # Override the defaults with the client node's attributes.
-  newnode['fqdn'].gsub!('_', '.') # Fix the fqdn, since periods couldn't be used in the databag ID.
-
-  nodes << newnode # Add the new node to the set of nodes.
-end
-
-# Filter out clients not in our environment, if applicable.
-if node['rdiff-backup']['server']['restrict-to-own-environment']
-  nodes.each do |n|
-    if n['chef_environment'] != node.chef_environment
-      nodes.delete(n)
-    end
+  # Copy over the check_rdiff nrpe plugin.
+  cookbook_file "#{servernode['rdiff-backup']['server']['nagios']['plugin-dir']}/check_rdiff" do
+    source "nagios/plugins/check_rdiff"
+    mode '755'
+    action :create
   end
+
+  # Give the user sudo access for the nrpe plugin.
+  sudo 'nrpe' do
+    user      'nrpe'
+    runas     'root'
+    nopasswd  true
+    commands  ["#{node['nagios']['plugin_dir']}/check_rdiff"]
+  end
+
 end
 
-if nodes.empty?
-  Chef::Log.info("WARNING: No nodes returned from search or found in rdiff-backup_unmanagedclients databag.")
-end
+puts "DEBUG: servernode"
+pp(servernode)
+
+# Note: The server backup user's private key must be copied over manually.
 
 jobs = Set.new
 
 # For each node, create a new job object for each job by merging job-specific attributes over node-specific ones.
 nodes.each do |n|
-  n['rdiff-backup']['client']['jobs'].keys.each do |src-dir|
-    if src-dir.start_with?("/") # Only work with absolute paths. Also excludes the "default" hash.
-      job = n['rdiff-backup']['client']['jobs']['default'] || Hash.new # Start with the client's default attributes if they are specified.
-      deep_merge!(job, n['rdiff-backup']['client']['jobs'][src-dir]) # Merge the job-specific attributes over the top.
+  if n['rdiff-backup']['client'].empty?
+    Chef::Log.info("WARNING: No jobs specified for host '#{n['fqdn']}'.")
+  end
+  n['rdiff-backup']['client'].keys.each do |src|
+    if src.start_with?("/") # Only work with absolute paths. Also excludes the "default" hash.
+      job = deep_copy(n['rdiff-backup']['client']['default']) || Hash.new # Start with the client's default attributes if they are specified.
+      deep_merge!(job, n['rdiff-backup']['client'][src]) # Merge the job-specific attributes over the top.
       job['fqdn'] = n['fqdn'] # Keep the fqdn in the job object.
-      job['source-dir'] = src-dir # Keep the source-dir in the job object.
-      jobs << job # Add the new job to the set of jobs.
+      job['source-dir'] = src # Keep the source-dir in the job object.
+
+      # Remove exclusion rules that don't apply to this job
+      relevantdirs = Array.new
+      job['exclude-dirs'].each do |dir|
+        relevantdirs << dir if dir.start_with?(src)
+      end
+      job['exclude-dirs'] = relevantdirs
+
+      jobs << job
     end
   end
 end
@@ -103,43 +195,53 @@ specifiedjobs = Set.new
 jobs.each do |job|
   newjob = Hash.new # Create a new "bare" job with just enough information to identify it.
   newjob['fqdn'] = job['fqdn']
-  newjob['source-dir'] = job['source-dir']
-  specifiedjobs << newjob # Add the job name of the form "fqdn dir" to the set of existing jobs.
+  newjob['source-dir'] = job['source-dir'].gsub("/", "-")
+  specifiedjobs << newjob
 end
+
+puts "DEBUG: specifiedjobs"
+pp(specifiedjobs)
 
 # Get the set of jobs which already exist so we can decide which ones to remove.
 existingjobs = Set.new
-File.open(CRON_FILE, "r") do |file|
-  file.each_line do |line|
-    if line.match(/^\D.*/) == nil # Only parse lines that start with numbers, i.e. actual jobs.
-      newjob = Hash.new # Create a new "bare" job with just enough information to identify it.
-      newjob['fqdn'] = line.split[7]
-      newjob['source-dir'] = line.split[8]
-      existingjobs << newjob # Add the job name of the form "fqdn dir" to the set of existing jobs.
+if File.exists?(CRON_FILE)
+  File.open(CRON_FILE, "r") do |file|
+    file.each_line do |line|
+      if line.match(/^\D.*/) == nil # Only parse lines that start with numbers, i.e. actual jobs.
+        newjob = Hash.new # Create a new "bare" job with just enough information to identify it.
+        newjob['fqdn'] = line.gsub(/.*\/(.*)_.*/, '\1').strip
+        newjob['source-dir'] = line.split('_')[-1].strip
+        existingjobs << newjob
+      end
     end
   end
 end
+
+puts "DEBUG: existingjobs"
+pp(existingjobs)
 
 # Get the set of jobs that need to be removed by subtracting the set of specified jobs by the set of existing jobs.
 removejobs = existingjobs.dup.subtract(specifiedjobs)
 
 # Figure out how much time to wait between starting jobs.
-minutesbetweenjobs = ((node['rdiff-backup']['server']['end-hour'] - node['rdiff-backup']['server']['start-hour'] + 24) % 24 * 60 ) / jobs.size
-hoursbetweenjobs = minutesbetweenjobs / 60
+unless jobs.empty?
+  minutesbetweenjobs = ((servernode['rdiff-backup']['server']['end-hour'] - servernode['rdiff-backup']['server']['start-hour'] + 24) % 24 * 60 ) / jobs.size
+end
 
 # Set up each job.
 setjobs = 0
 jobs.each do |job|
 
+  puts "DEBUG: job"
+  pp(job)
+
   # Shorten some long variables for readability.
   fqdn = job['fqdn']
   sd = job['source-dir']
   dd = "#{job['destination-dir']}/filesystem/#{fqdn}#{sd}"
-  suser = node['rdiff-backup']['server']['user']
+  suser = servernode['rdiff-backup']['server']['user']
   maxchange = job['nagios']['max-change']
   latestart = job['nagios']['max-late-start']
-  latefinwarn = job['hour'] + (job['minute']+59)/60 + job['nagios']['max-late-finish-warning'] # Minute is ceiling'd up to the next hour
-  latefincrit = job['hour'] + (job['minute']+59)/60 + job['nagios']['max-late-finish-critical'] # Minute is ceiling'd up to the next hour
   servicename = "rdiff-backup_#{job['fqdn']}_#{sd}"
   nrpecheckname = "check_rdiff-backup_#{job['fqdn']}_#{sd.gsub("/", "-")}"
 
@@ -154,50 +256,69 @@ jobs.each do |job|
 
   # Set run times for each job, distributing them evenly across a certain time period every day.
   job['minute'] = (minutesbetweenjobs * setjobs) % 60
-  job['hour'] = ((hoursbetweenjobs * setjobs) + node['rdiff-backup']['server']['start-hour']) % 24
+  job['hour'] = ((minutesbetweenjobs * setjobs / 60).floor + servernode['rdiff-backup']['server']['start-hour']) % 24
   setjobs += 1
 
   # Create the exclude files for each job.
-  template "/home/#{suser}/exclude/#{fqdn}-#{sd}" do
-    source "exclude.sh.erb"
-    mode "0774"
+  directory "/home/#{suser}/exclude" do
+    owner suser
+    group suser
+    mode '0775'
+    recursive true
+    action :create
+  end
+  template "/home/#{suser}/exclude/#{fqdn}_#{sd.gsub("/", "-")}" do
+    source "exclude.erb"
+    owner suser
+    group suser
+    mode "0664"
     variables({
-        :fqdn => fqdn,
-        :src => sd,
-        :paths => job['exclude-dirs'],
-      })
+      :src => sd,
+      :paths => job['exclude-dirs'],
+    })
     action :create
   end
 
   # Create scripts for each job.
-  template "/home/#{suser}/scripts/#{fqdn}-#{sd}" do
+  directory "/home/#{suser}/scripts" do
+    owner suser
+    group suser
+    mode '0775'
+    recursive true
+    action :create
+  end
+  template "/home/#{suser}/scripts/#{fqdn}_#{sd.gsub("/", "-")}" do
     source "job.sh.erb"
-    mode "0664"
+    owner suser
+    group suser
+    mode "0774"
     variables({
-        :fqdn => fqdn,
-        :src => sd,
-        :dest => dd,
-        :period => job['retention-period'],
-        :user => job['user'],
-        :port => job['ssh-port'],
-        :args => job['additional-args']
-      })
+      :fqdn => fqdn,
+      :src => sd,
+      :dest => dd,
+      :period => job['retention-period'],
+      :user => job['user'],
+      :port => job['ssh-port'],
+      :args => job['additional-args']
+    })
     action :create
   end
   
   # If nagios alerts are enabled and the job didn't already exist, create nagios alerts for the job.
-  if node['rdiff-backup']['server']['nagios']['alerts'] and job['nagios']['alerts']
-    if existingjobs.include?("#{fqdn} #{sd}")
+  if servernode['rdiff-backup']['server']['nagios']['alerts'] and job['nagios']['alerts'] and not existingjobs.include?("#{fqdn} #{sd}")
 
     # TODO: But what about if nagios attributes get updated? How do we detect it? Maybe we should just always run the lwrps to update them? Or do we need to delete them first? Won't that reset the alert so it has no history?
 
+    latefinwarn = job['hour'] + (job['minute']+59)/60 + job['nagios']['max-late-finish-warning'] # Minute is ceiling'd up to the next hour
+    latefincrit = job['hour'] + (job['minute']+59)/60 + job['nagios']['max-late-finish-critical'] # Minute is ceiling'd up to the next hour
+
     nagios_service servicename do
       command_line "$USER1$/check_nrpe -H $HOSTADDRESS$ -c #{nrpecheckname}"
-      host_name node['fqdn']
+      host_name servernode['fqdn']
       action :add
     end
     nagios_nrpecheck nrpecheckname do
-      command "sudo #{node['nagios']['plugin_dir']}/check_rdiff -r #{dd} -w #{latefinwarn} -c #{latefincrit} -l #{maxchange} -p #{latestart}"
+      command "sudo #{servernode['rdiff-backup']['server']['nagios']['plugin-dir']}/check_rdiff -r #{dd} -w #{latefinwarn} -c #{latefincrit} -l #{maxchange} -p #{latestart}"
       action :add
     end
   else # Remove the nagios alerts if alerts are disabled.
@@ -215,44 +336,26 @@ template CRON_FILE do
   source "cron.d.erb"
   mode "0664"
   variables({
-      :fqdn => job['fqdn'],
-      :src => job['source-dir'],
-      :dest => job['destination-dir'],
-      :period => job['retention-period'],
-      :user => job['user'],
-      :port => job['ssh-port'],
-      :args => job['additional-args']
-    })
+    :shour => servernode['rdiff-backup']['server']['start-hour'],
+    :ehour => servernode['rdiff-backup']['server']['end-hour'],
+    :mailto => servernode['rdiff-backup']['server']['mailto'],
+    :suser => servernode['rdiff-backup']['server']['user'],
+    :jobs => jobs
+  })
   action :create
 end
 
+puts "DEBUG: removejobs"
+pp(removejobs)
+
 # Remove the exclude files, job scripts, and nagios alerts for each job to be removed.
 removejobs.each do |job|
-  File.delete("/home/#{node['rdiff-backup']['server']['user']}/exclude/#{job['fqdn']}-#{job['source-dir']}")
-  File.delete("/home/#{node['rdiff-backup']['server']['user']}/scripts/#{job['fqdn']}-#{job['source-dir']}")
+  File.delete("/home/#{servernode['rdiff-backup']['server']['user']}/exclude/#{job['fqdn']}_#{job['source-dir']}")
+  File.delete("/home/#{servernode['rdiff-backup']['server']['user']}/scripts/#{job['fqdn']}_#{job['source-dir']}")
   nagios_nrpecheck "check_rdiff-backup_#{job['fqdn']}_#{job['source-dir'].gsub("/", "-")}" do
     action :remove
   end
   nagios_service "rdiff-backup_#{job['fqdn']}_#{job['source-dir']}" do
     action :remove
-  end
-end
-
-# Recursive copy
-def deep_copy(object)
-  return Marshal.load(Marshal.dump(object))
-end
-
-# Recursive bot.merge!(top) for hashes.
-def deep_merge!(bot, top)
-  top.keys.each do |key|
-    if top[key].is_a?(Hash)
-      if !bot[key].is_a?(Hash) # Make a new subhash in bot if top has one there.
-        bot[key] = Hash.new
-      end
-      deep_merge!(bot[key], top[key])
-    else
-      bot[key] = top[key]
-    end
   end
 end
